blacklist and auth edit group

+++ midterm/server/routes/group/groupsController.js
@@ -1,4 +1,4 @@
-const { Groups, Account } = require('../../models');
+const { Groups, User } = require('../../models');
 
 async function getGroup() {
   try {
@@ -30,7 +30,8 @@ async function createGroup({ groupname, members }) {
     if (!groupname) {
       return { status: false, message: 'Invalid Infomation!' };
     }
-    const existsMember = await Account.find({ id: { $in: members } }).lean();
+    // Kiểm tra member nhập vào có trong bảng user hay không
+    const existsMember = await User.find({ id: { $in: members } }).lean();
     if (!existsMember) {
       return { status: false, message: 'Invalid Infomation!' };
     }
@@ -47,27 +48,46 @@ async function createGroup({ groupname, members }) {
     throw error;
   }
 }
-async function editGroup(groupId, groupInfo) {
+async function editGroup(userInfo, groupId, groupInfo) {
   try {
     if (!groupId) {
       return { status: false, message: 'Invalid Infomation!' };
     }
-
-    const existsMember = await Account.find({ _id: groupInfo.members }).lean();
+    // Kiểm tra có phải owner hay không?
+    const IsOwner = await Groups.findOne({ owner: userInfo.id }).lean();
+    if (!IsOwner) {
+      return { status: false, message: 'Invalid Credenticals!' };
+    }
+    //Kiểm tra xem có là co-owner hay member chưa?
+    const existedUser = await Groups.findOne({
+      $or: [{ members: groupInfo.members }, { coowner: groupInfo.coowner }],
+    }).lean();
+    if (existedUser) {
+      return { status: false, message: 'Invalid Credenticals!' };
+    }
+    //Kiểm tra id nhập vào có trong bảng user hay không?
+    const existsMember = await User.find({ _id: groupInfo.members }).lean();
     if (!existsMember) {
       return { status: false, message: 'Invalid Infomation!' };
     }
+    //Update sau khi đã kiểm tra
     const myGroups = await Groups.findOneAndUpdate({ groupId }, { groupInfo });
     return { status: true, message: 'update successful!', myGroups };
   } catch (error) {
     throw error;
   }
 }
-async function deleteGroup({ groupId }) {
+async function deleteGroup(userInfo, { groupId }) {
   try {
     if (!groupId) {
       return { status: false, message: 'Invalid Infomation!' };
     }
+    // Kiểm tra có phải owner hay không?
+    const IsOwner = await Groups.findOne({ owner: userInfo.id }).lean();
+    if (!IsOwner) {
+      return { status: false, message: 'Invalid Credenticals!' };
+    }
+
     await Groups.findOneAndDelete({ groupId });
     return { status: true, message: 'delete successful!' };
   } catch (error) {
